# backend/openapi.yml
openapi: 3.0.3
info:
  title: Kickstat API
  description: Sport Coaching Tool API for managing squads, athletes, events, and live match logging.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development
paths:
...
  /api/athletes:
    get:
      summary: List all athletes in the squad
      tags: [Athletes]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of athletes
    post:
      summary: Add an athlete (coach only)
      tags: [Athletes]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                position: { type: string }
                squad_number: { type: integer }
                date_of_birth: { type: string, format: date }
                contact_info: { type: string }
      responses:
        '201':
          description: Athlete created
        '403':
          description: Not authorized (assistant role)
  /api/athletes/{id}:
    patch:
      summary: Update an athlete (coach only)
      tags: [Athletes]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Updated athlete
    delete:
      summary: Remove an athlete (coach only)
      tags: [Athletes]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: Athlete removed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Clerk JWT token